+++ src/routers/mod.rs
@@ -3,6 +3,7 @@ mod bulk_router_message;
 mod remove_actor_message;
 mod round_robin_router;
 mod router_message;
+mod sharded_router;
 
 pub mod prelude {
     pub use crate::routers::add_actor_message::AddActorMessage;
@@ -11,4 +12,8 @@ pub mod prelude {
     pub use crate::routers::round_robin_router::RoundRobinRouterFactory;
     pub use crate::routers::round_robin_router::RoundRobinRouter;
     pub use crate::routers::router_message::RouterMessage;
+    pub use crate::routers::sharded_router::ShardedRouter;
+    pub use crate::routers::sharded_router::ShardedRouterFactory;
+
+
 }